requirements.txt
# Discord Bot Dependencies

# Core Discord
discord.py==2.3.2
discord.py[voice]==2.3.2
PyNaCl==1.5.0

# AI & API
groq>=0.4.0
openai>=1.10.0
python-dotenv==1.0.0

# Music & Media
yt-dlp>=2024.03.10
pydub>=0.25.1

# Web & API
flask==3.1.3
gunicorn==25.2.0
requests>=2.31.0
werkzeug>=3.1.0
jinja2>=3.1.2

# Utilities
python-dateutil>=2.8.0
aiohttp>=3.8.0
